Implementacion de seeders para la carga inicial de datos

+++ src/main/java/com/example/back_end/model/Plato.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "platos")
 @Data
@@ -21,7 +23,7 @@ public class Plato {
     private String nombre;
 
     @Column(nullable = false)
-    private Double precio;
+    private BigDecimal precio;
 
     @Column(length = 255)
     private String descripcion;
